GPU fallback and XML
Here is the changes I did to AdapterOptions.xml and Driver.cpp, To read settings xml and insert it into AdapterOptions.H from xmlvalue.
if xml is not found, search for adapter.txt.

if wrong GPU name in either, autoselect the best one (vram based)

loadsettings is meant to replace loadoptions, but is missing alot of code since ths is the order you asked me to present it. So no fallbck for missing xml or options.tx in this pr