Better parsing and limiting and sorting of videomodes
Limiting the total modes to 92, had issues with to many modes in vddd_settings.xml hindering initializing of VirtualDisplay. When we limit we cut the mode lines over 92, so inorder to let the user chose which modlines to cut we have a sort order. x-desc | x-ass | y-desc| y-ass | ref.rate-desc | ref.rate-ass

This is set in vdd_settings.xml, and if the xml is missing the correct tag it will default to x-desc.

We also makesure there are no duplicate modelines from edid, like td 1024*768@60 and a identical VIC in CEA block.