Removed most fail points
Now it will take "Friendlyname, 3" for PCI-bus 3, Device 0, Function 0. If Frienlyname alone it will match with the first, else as allways, bestGpu(most vram) selected.

Now it compiles again, I had to up the C++ version to 17 in order to get "optional" working. There are probably tricks for it, but i don't have time or need to put effort into this.
DevID and iDX stuff remove due to Kernel/Usermode not playing nicely.

Will see if i up a temporaary  ".snl and/or .vxproj.user" for simplicity to build