Changed version of C++ 17
WE have used v.14 along time to be backwards compatible, but now that we have stepped up to udf 2.25 and iddcx 1.0a (last win10 version support), I've chose to step it up due to "ease of use", optional can be hacked around into v.14 code, but full support in 17. And since all other code works in 17, I see it as "future proving"

More changes to project file might be needed since i can test it  while VS2022 is reinstalling. But I had issues with VS not recognizing and allowing UDFM mode but handle it all as an .dll