Still Code 31
Compiles, but get Code31 at install. Need feedback to proceed. I thought it might be a missing "default handler" which got lost in some of the 3 trial versions I've  tried. But this last commit has a default/fallback handler again. So that's not the reason for Code31.

Running out of ideas. It should work in theory.